White theme for transparent pages
If the element at the top is transparent, theme color should be set to white.

Before the update, transparent background would cause rgb(0, 0, 0) theme color (black), while the browser would render transparent background as white. This would create a mismatch between the page and browser theme. The problem was spotted on skelbiu.lt - the change may be tested there.